Refactor WriteReview block into its own component
